requirements.txt
Pillow
pyttsx3
pyautogui
pytesseract
numpy
playsound
pydub
gtts
pygame
wave
simpleaudio
